+++ packages/signals/lib/signals_flutter.dart
@@ -4,4 +4,6 @@ export 'signals.dart';
 export 'src/extensions/signal.dart';
 export 'src/extensions/value_listenable.dart';
 export 'src/extensions/value_notifier.dart';
+export 'src/widgets/watch.dart';
+export 'src/widgets/signal_widget.dart';
 export 'src/watch_signal.dart';